[PORT FROM MAIN] Fix the timestamp overflow bug
BZ: 28089 50102

This patch fix an Android defeat. 32 bits timestamp is not big enough,
change to 64 bits to avoid overflow.

Original patch:
R3: 41774

Orig-Change-Id: Ie3c5a7f485e8a31949a9c4c979a97eae704e9fa3

+++ media/libstagefright/MPEG4Extractor.cpp
@@ -446,7 +446,7 @@ sp<MetaData> MPEG4Extractor::getTrackMetaData(
                 }
             } else {
                 uint32_t sampleIndex;
-                uint32_t sampleTime;
+                int64_t sampleTime;
                 if (track->sampleTable->findThumbnailSample(&sampleIndex) == OK
                         && track->sampleTable->getMetaDataForSample(
                             sampleIndex, NULL /* offset */, NULL /* size */,
@@ -3115,7 +3115,7 @@ status_t MPEG4Source::read(
                     sampleIndex, &syncSampleIndex, findFlags);
         }
 
-        uint32_t sampleTime;
+        int64_t sampleTime;
         if (err == OK) {
             err = mSampleTable->getMetaDataForSample(
                     sampleIndex, NULL, NULL, &sampleTime);
@@ -3161,7 +3161,7 @@ status_t MPEG4Source::read(
 
     off64_t offset;
     size_t size;
-    uint32_t cts;
+    int64_t cts;
     bool isSyncSample;
     bool newBuffer = false;
     if (mBuffer == NULL) {
